pidgeon/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
decorator==4.4.2
docopt==0.6.2
greenlet==0.4.13
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pinject==0.14.1
PyNaCl==1.4.0
PyYAML==5.3.1
readline==6.2.4.1
six==1.15.0
typing-extensions==3.7.4.2
yarl==1.4.2
